Unify check engine specs
