+++ tests/Authorizee.Core.Tests/Authorizee.Core.Tests.csproj
@@ -13,7 +13,7 @@
         <PackageReference Include="FluentAssertions" Version="6.12.0" />
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.6.0"/>
         <PackageReference Include="NSubstitute" Version="5.1.0" />
-        <PackageReference Include="Verify.Xunit" Version="23.2.1" />
+        <PackageReference Include="Verify.Xunit" Version="23.2.2" />
         <PackageReference Include="xunit" Version="2.7.0" />
         <PackageReference Include="xunit.runner.visualstudio" Version="2.5.7">
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
