--- tests/Authorizee.Data.Postgres.Tests/CheckEngineSpecs.SubjectPermissionShouldEvaluatePermissions.verified.txt
@@ -1,52 +1,0 @@
-﻿{
-  permission_0: true,
-  permission_1: true,
-  permission_10: true,
-  permission_11: true,
-  permission_12: true,
-  permission_13: true,
-  permission_14: true,
-  permission_15: true,
-  permission_16: true,
-  permission_17: true,
-  permission_18: true,
-  permission_19: true,
-  permission_2: true,
-  permission_20: true,
-  permission_21: true,
-  permission_22: true,
-  permission_23: true,
-  permission_24: true,
-  permission_25: true,
-  permission_26: true,
-  permission_27: true,
-  permission_28: true,
-  permission_29: true,
-  permission_3: true,
-  permission_30: true,
-  permission_31: true,
-  permission_32: true,
-  permission_33: true,
-  permission_34: true,
-  permission_35: true,
-  permission_36: true,
-  permission_37: true,
-  permission_38: true,
-  permission_39: true,
-  permission_4: true,
-  permission_40: true,
-  permission_41: true,
-  permission_42: true,
-  permission_43: true,
-  permission_44: true,
-  permission_45: true,
-  permission_46: true,
-  permission_47: true,
-  permission_48: true,
-  permission_49: true,
-  permission_5: true,
-  permission_6: true,
-  permission_7: true,
-  permission_8: true,
-  permission_9: true
-}
