--- tests/Authorizee.Data.Postgres.Tests/CheckEngineSpecs.SubjectPermissionShouldListAllPermissions.verified.txt
@@ -1,52 +1,0 @@
-﻿{
-  permission_0: false,
-  permission_1: false,
-  permission_10: false,
-  permission_11: false,
-  permission_12: false,
-  permission_13: false,
-  permission_14: false,
-  permission_15: false,
-  permission_16: false,
-  permission_17: false,
-  permission_18: false,
-  permission_19: false,
-  permission_2: false,
-  permission_20: false,
-  permission_21: false,
-  permission_22: false,
-  permission_23: false,
-  permission_24: false,
-  permission_25: false,
-  permission_26: false,
-  permission_27: false,
-  permission_28: false,
-  permission_29: false,
-  permission_3: false,
-  permission_30: false,
-  permission_31: false,
-  permission_32: false,
-  permission_33: false,
-  permission_34: false,
-  permission_35: false,
-  permission_36: false,
-  permission_37: false,
-  permission_38: false,
-  permission_39: false,
-  permission_4: false,
-  permission_40: false,
-  permission_41: false,
-  permission_42: false,
-  permission_43: false,
-  permission_44: false,
-  permission_45: false,
-  permission_46: false,
-  permission_47: false,
-  permission_48: false,
-  permission_49: false,
-  permission_5: false,
-  permission_6: false,
-  permission_7: false,
-  permission_8: false,
-  permission_9: false
-}
