+++ tests/Authorizee.Tests.Shared/Authorizee.Tests.Shared.csproj
@@ -7,6 +7,8 @@
     </PropertyGroup>
 
     <ItemGroup>
+      <PackageReference Include="FluentAssertions" Version="6.12.0" />
+      <PackageReference Include="Verify.Xunit" Version="23.2.2" />
       <PackageReference Include="xunit" Version="2.7.0" />
     </ItemGroup>
